Added site variables ('sitevars'). These are stored in the sitevariables database table. Each one has a name, a description (NOT OPTIONAL!), a default value, and a current value. If the current value is NULL, the default value is used.
Soon, a mechanism will be added to the install process to
ensure all needed site variables exist before installing;
more on that when it is committed.

 - Added 'editsitevars.php3' page, accessable to admins
   via the 'Edit Site Variables' menu option.

 - Added 'setsitevar' script,
   an interface for listing, viewing in detail, and setting
   site variables.

 - Web interface now uses 'web/nologins' and 'web/message'
   instead of one-off database tables.

NOTE that setting a variable to the default value and
setting a variable to a value which is string-identical
are NOT the same thing.
(This doesn't matter yet, but when we push default values out to
 remote sites as part of our install, it will.)