Okay, try again with correct comment: Switch from MD5 verification of all files (fdisk, slice images, tarballs) to proper digital signatures using the new priv/pub key pair recently installed (and the mksig script). The key pair is password protected, so it is not as convenient as an MD5, but such is life. The sig files are either stored locally on the CD, or can be specifed as an https:// URL to be downloaded with wget and verified using the public key stored on the CD in /etc/emulab_pubkey.pem.
This version of the script is not compatible with older CDs (version 3
or earlier). The web interface will return the proper version to the
requester based on its CD version (from boss:/z/testbed/distributions).
The old version is stored as netbed-setup-v3.pl, while the new version
is stored as netbed-setup-v4.pl. These names are relevant to the web
interface (cdromcheckin.php3), but are otherwise arbitrary (they can
be anything). Note that version 1 and 2 CDs do not download a script
(hardwired on the CD).