More bloat! Check for experiments in transition. Check state for active. Add explanation of how to use the page at the top. Add check box for modifying the base experiment (virt_lans). Redo Red/Gred stuff so that its "greyed" out on links that are not red/gred. Clean up the script argument processing and proper permission checks.
The missing feature that would be nice to add is the ability to modify
the base experiment config when the experiment is swapped out. I think
it would be very handy. The problem is that instead of looking at
delays and linkdelays, we have to read virt_lans directly and do a
bunch of stuff that assign_wrapper does when creating delays and
linkdelays; namely putting together the individual member parameters
of a link or lan into a coherent set of values. Not rocket science,
but a lot of ugly script hacking, and I am not in the mood. Note
though that I modified the backend (delay_config) to allow this to be
done in the future.