Add mechanism to allow admins to log in even when web interface is turned off. Use this link directly:
	https://www.emulab.net/login.php3?adminmode=1

This is passed into DOLOGIN(), and if the user is a real admin (of
course) the login is created with the adminmode set to on instead of
off (the default). This essentially causes the NOLOGINS checks later
on to be bypassed (as is the case when its an admin that is logged
in).

I prefer this approach cause it hides the entire admin thing, rather
than putting up a checkbox that might leak out to a mere user (remote
chance, but I feel better this way).

There is also some new stats code that is turned off until its
finished.