Add several minor experiment_stats fields; swap_errors (a count), swap_exitcode (last error), idle_swaps (a count), batch (a flag to indicate a batch experiment).
Add a operational log. Okay, its not actually a log, but a table that
will grow forever until it consumes the earth. Its a small table
though, so it will take a few years. Its cross indexed with the
experiment_stats table, so by massaging this table along with the
stats table, we can get a good picture of what was running on the
testbed when, and how many resources it was using. Sorry, not a log
file, but we can easily generate a log file from tbe table if the Boss
really wants one. The table entry averages 28 bytes.

Move stats to their own main menu item (admin mode only). Remove from
the showexp_list page since that was bogus.