Fix a nit and a long-standing bug.
If I was red dot when my login timed out (or whatever) the login link
should go back to admin mode after I log in.

The bug was that by the time I get into loggedinordie, if my login isn't
valid, getuid fails, and uid is "", instead of what the cookie said. This
caused it always just to say "You do not appear to be logged in!" instead
of the real error (ie timed out, etc). So if uid=="", grab the uid from
the cookie instead of checking things for nobody.