Show staleness of idle data in more places, since the only place it showed before was idle view of the expt list. Now it shows in all (non-thumb) views.
I also added it to the right hand table in the expt page.

It also shows in the little expt lists in the user, group, and proj pages.
(I'm already really glad those got consolidated into a single function.)

Also cleaned up idle view a bit. Now that we show staleness and idle
ignore anyway, we don't need text tags for them, and we no longer need the
red dot that links to swap requests because of idlemail. So there's no
longer the weird double-column thing in idle view.

I'm still looking for a good way to non-intrusively mark expts in the expt
list as being unswappable, so if you think of anything, let me know.
(Maybe boldface, or a little lock icon or something?)