Fix a slight race condition that has already bitten us several times in real life.
There's a small window in time (typically a few seconds, but maybe 10's of
seconds on occasion) between the time nodes get nalloc'd and the time they
get rebooted during a swapin. If idlemail runs (or someone views a web
page) during this time, the expt can appear idle because the nodes haven't
had any activity. It can cause an incorrect email or a false positive on
the web page.

The fix is to have nalloc register some external node activity, like
node_reboot does, when it allocates a node.