Minor changes.
dojailconfig: Do not return a jail config unless the node is really
jailed.

doifconfig: Look at the node name, and return ifconfig info specific
to the name instead of always returning all info for the machine. Upon
boot, virtual interfaces will not be setup, but instead will be setup
when the jail is setup. When the jail requests ifconfig info, it gets
only its list of interfaces to setup.

dolinkdelays: Ditto, since linkdelays will be setup after per-jail
interfaces are setup.