Minor stats changes for dealing with swapmodify; be sure to credit for each portion of the experiment as it is modified.
Also add expt_swap_uid so that we know who did the last operation, and
so we can charge/credit the right person. So, if joe swaps in the
experiment and jane swaps it out, joe gets charged. If jane swaps in
the experiment and joe modifies it, jane gets credit for the first
portion, and joe will later get charged for the second portion.

Took longer to explain then to implement ...

Lbs