Round of changes related to idleswapping and autoswapping. The web and back end scripts now support 3 different kind of forced swaps:
1. Idle-Swap : this is ths same one we had before. Email message to them
says it was swapped "because it was idle for too long"

2. Auto-Swap : A new one, typically for user-requested timed swapouts.
Email says it was swapped "because it was swapped in too long"

3. Force swap: Generic one, for "none of the above" cases. Just says
Experiment "has been forcibly swapped out by Testbed Operations."

The force swap option on the web now lets you choose which of these three
you want. Only "Idle-Swap" counts as an idleswap in the stats. Soon
idleswap and autoswap will be used by idlemail when it does automatic
swapping.