Add the new fields related to our new swapping features:
   noswap_reason tinytext			- The reason they're unswappable
   idleswap tinyint(4) NOT NULL default '0'	- The idleswap bit (for auto. idleswap)
   idleswap_timeout int(4) NOT NULL default '0'	- The timeout for idleswap (in minutes)
   noidleswap_reason tinytext			- The reason they're non-idleswap
   autoswap tinyint(4) NOT NULL default '0'	- The autoswap bit
   autoswap_timeout int(4) NOT NULL default '0'	- The timeout for autoswap (in minutes)

The new form for expt create should fill these in, and swapin will update the idleswap
timeout if idleswap is set. Idlemail will read these and do the deed as needed.
They'll also get shown on the new expt display page.

As soon as we move to the new form, we can nuke priority.