Added the TRIVIAL_LINK_BW option, which allows, in the ptop file, specification of how much bandwidth can be used on the node for trivial links. Clearly, there will be some limit to loopback transfers, and this makes assign aware of it.
While I was in there, cleaned up and commented some of the code I was
working on.

Also fixed a bug with the type pre-check and emulated links that are
trivial_ok - we can't add these into the total for a vnode, because
they could end up being satisified with trivial links.