Minor changes to ifconfig and linkdelay to allow a veth interface to stand alone (no underlying phys interface) so that both ends of a virtual link can be on the same node.
Change "startcmd" to return the uid of the swapper, not the creator,
as the UID to run the startup command.

Get creator/swapper uid in the initial query for the node, instead of
as separate queries later. Saves some queries ...

Remove old code.

Add progagents command to return the list of program object names, and
the UID to run those objects as (again, the swapper not the creator).
This allows the program-agent startup code on the node to be more
selective in its elvin subscription, as well as running the programs
as the proper UID.

Changes to jailconfig. If the node has its own IP in the nodes table
(jailip), return that as JAILIP in the configuration for mkjail.