A fix for Mike's "Pointless Nit o' the Day" for today:
Add a magical icon for the Emulab pages. Shows up on the tab for tabbed browsers, and
in the bookmarks, etc.

The current image isn't that great, but it's something. When we've got time, we'll
make it a little better by not trying to fit so much of the logo in at once.