Hack for dealing with user output from the parser (puts). Since I spit the XML to stdout, user calls to puts confuse the converter, and get lost so the user never sees them (before, they would end up in the start/preload log). The converter now eats up lines till it sees a valid XML header string, printing those lines to stdout so they end up in the log like before.
Need to deal with this in a better manner at some point. I did not
want to write the XML to temporary file. Not sure why, but I thought
that was the right approach, to keep the user from getting his grubby
little hands on it. I'm thinking now thats silly, and I'll eventually
fix things up (since it requires more TCL hacking, and I don't have
the energy for that).