Split the -e option to ptopgen into -p/-e. Assign will always pass in the pid so that ptopgen can do permission checks on the node types and classes before it sticks them into the ptop file. -e now takes just eid, but operates as before.
Change the -v option slightly; assign passes -v option when it sees
that the topology requires virtual nodes. Without -v, virtual nodes
are not placed into the the ptop file, saving about 6000 lines of node
entries!

Get rid of the pcvm and pc601 hacks, and replace with a permission
check, as determined by the nodetypeXpid_permissions. I've updated
that table to reflect current types/classes.

Change Rob's last change, which was doing a DB query on each node,
which when used with -v, was issuing a lot of queries. I was puzzled
why ptopgen was taking 4 seconds to run!

Kill more sharks.
Cleanup some terrible indenting.