Fix a major bug in dynamic pclasses. One of thse little things that was sooo wrong you wonder how it worked in the first place... Took me days to find this one!
Also added a new switch, '-o', that lets assign try out solutions
that over-load a pnode. This helps a lot with topologies where the
optimal solution is a best-fit onto multiplexed pnodes.

The end result is that Mike's snake maps much better - it used to get
an essentially random mapping, but now it gets something acceptible.