Wow. Mountain of changes for the new begin experiment form.
Lots of changes to the form, both functional and aesthetic. See the
testbed ops mail logs for a list of all of them, and the rationale.

Corresponding updates to the showexp "edit meta-data" stuff, so that it
gets all the same error checks as the real form.

Also some backend changes in batchexp to pass through all the new form
values.