Clean up temporary files used in modify. The temp dirs were being created in /tmp and left behind. I've moved them to the expwork directory instead, and added a routine in the library to clear them out.
Clear out the nsfile (stored in /tmp) used in modify. The web page was
creating a temp file, but never removing it. swapexp now copies the
nsfile in so that the web page can remove the temporary after the
script exits. The temp is placed in the expwork directory as well, but
left behind for debugging.

When swapmod fails, send along the nsfile in the email message.