A lot of bugfixes relating, primarily, to multiplexed nodes:
Split the function that determines if a vnode can be mapped to a given pnode
out of find_pnode, so that I can spread it out and make it more comprehensible.
In doing so, fixed a bug in which static types were not getting handled
correctly - a static type would only be found if no vnodes with dynamic types
were allocated to the pnode.

Fixed a bug in which, on allocation of a dynamic ptype, nodes where being
removed from their static type lists in the type_table.

Fixed a bug when there is a combination of fixed and un-fixed vnodes in the top
file.

I have more on the way, but this seemed like a good checkpoint.