(Most of) the rest of the gated-in-jail changes.
Add a -V option to mkjail.pl telling it to use the virtual control net
IP (172.16.0.1) for the default route instead of the real control net
IP (155.101.132.1).  In this case, it also assigns a virtual control net
IP alias (172.17.<pnode>.0) to the physical interface so that the
aforementioned routes can actually be installed.

Add the same -V option (only valid with -j) in vnodesetup so that we
can get it through from bootvnodes.  (Also cleaned up the usage message
some)

The virtual control net address/mask are currently constants in mkjail.pl.
At some point they can/should be changed to config time options so that we
can easily use this on minibed too.

In theory, this (-V) shouldn't be an option and we should just use
virtual control net routing all the time.  But I want to get some more
testing and we have to resolve the minibed conflicts first.  So for now,
this is an option and it is off by default.

Haven't tied this option in with gated yet, so creating a vnode experiment
with Session routing still won't work.  If you want to do that, create the
experiment, watch gated blow up right and left, login to all the physical
nodes and change bootvnodes to add -V to the invocation of vnodesetup,
and then reboot everything.  Now it should be working!