Add some rather bogus code to close all the fd's (except 0,1,2) before execing the program. We had a problem with daemons that detach and continue to run, holding open all of the same FDs that the web server had open, including the httpd socket! So it was not possible to restart the web server cause processes were parked on the socket holding it open.
I think this is the wrong place for this since there numerous ways to
run programs from the web server, and using suexec is optional (can
just run programs directly). Seems like a problem with php instead,
but the fix is here cause it will solve the problem for us.