Widearea changes, mostly started for plab but affects all widearea nodes. The intent is to better support last mile types, which right now are a total mess cause the types are associated with the virtual nodes, and a node can have just a single type. This info has now been moved in the node_types_auxtypes table and the node_types table.
Anyway, we no longer use wanassign on unconnected widearea nodes, but
use assign directly, much like we use assign to allocate virtual nodes
on local physical nodes. ptopgen inserts the physical widearea nodes
(even though they are allocated), and the proper counts for the types
that are available on them. assign will pick the nodes, and
assign_wrapper will allocate the necessary vnodes on the pnodes (but
in the case of wideare nodes, the underlying physical node does not
need to be allocated).

Also added some fixes for dealing with vtypes when used in conjunction
with widearea nodes. Rather then generating an error like it used to,
you can create a vtype in your NS file:

	tb-make-soft-vtype mytype {pcvroninet pcvwainet}
	tb-set-hardware $v0 mytype

and assign_wrapper now looks at the underlying types to figure out
what it needs. Note: No consistency checking yet; mixing a remote/virt
and a local/real type will break.