Some fixes aimed at vnodes:
Don't allow intraswitch links to vnodes on switches - they'll get
scored as direct links. Since LANs are now directly on switches, there
were situations where we could have scored the same link as either
direct or intraswitch, which results in two different scores. This can
cause those 'contact calfeld' warnings, and might have caused some
strange scoring fluxuations as assign ran.

When picking a link to a switch, for emulated links, pick the least
loaded link, not the one with the fewest users. This helps in situations
where the link bandwidths are heterogenous.