Small change to libplab to have it try to pull the plab rootball from a little local loopback httpd server which now runs on all slivers of the static utah1 slice.
It falls back to pushing the tarball from Emulab central if the local
fetch fails.  Also introduced the notion of versioning for the plab rootball.
For now, just chang the ROOTBALL_NAME def at the beginning of this file; I'd
like to move this into autoconf or the makefile eventually.